refactor: Move OpenAPI-specific filters to a separate module
